+++ React/fast-company/src/components/qualitie.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export const Qualitie = ({color, name, _id}) => {
+
+  return <div style={{marginRight: '5px'}} className={`badge bg-${color}`} key={_id}>{name}</div>;
+
+}
