+++ React/fast-company/src/index.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import { Users } from './components/users';
 import reportWebVitals from './reportWebVitals';
+import { App } from './app';
 import 'bootstrap/dist/css/bootstrap.css';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    <Users />
+    <App />
   </React.StrictMode>
 );
